+++ doc/source/changelog.rst
@@ -3,6 +3,23 @@ Changelog
 #########
 
 **********************
+v0.2.1 (Nov 17, 2014)
+**********************
+
+Dependencies
+============
+
+- pandas version > 0.15. This required some change in the xls reading within
+  the parser.
+- pytest. For the unit tests.
+
+Misc
+====
+
+- Unit testing for all mathematical functions and a first system wide check.
+- Fixed some mistakes in the tutorials and readme
+
+**********************
 v0.2.0 (Sept 11, 2014)
 **********************